Fix bug that caused the app to crash when an existing project is search for.
Our project is setup a bit weird rn and we totally need to fix it. It's weird because the user has no choice over which options to choose from the search results. We need to provide them with a list of options and let them decide which project they want.